test: split redis instrumentation test
