+++ instrumentation/redis/CHANGELOG.md
@@ -1,18 +1,22 @@
 # Release History: opentelemetry-instrumentation-redis
 
+### Unreleased
+
+test: split redis instrumentation test [#754](https://github.com/open-telemetry/opentelemetry-ruby/pull/754)
+
 ### v0.17.0 / 2021-04-22
 
 * (No significant changes)
 
 ### v0.16.0 / 2021-03-17
 
-* FIXED: Update DB semantic conventions 
-* FIXED: Example scripts now reference local common lib 
-* DOCS: Replace Gitter with GitHub Discussions 
+* FIXED: Update DB semantic conventions
+* FIXED: Example scripts now reference local common lib
+* DOCS: Replace Gitter with GitHub Discussions
 
 ### v0.15.0 / 2021-02-18
 
-* ADDED: Add instrumentation config validation 
+* ADDED: Add instrumentation config validation
 
 ### v0.14.0 / 2021-02-03
 
@@ -28,13 +32,13 @@
 
 ### v0.11.0 / 2020-12-11
 
-* ADDED: Accept config for redis peer service attribute 
-* ADDED: Move utf8 encoding to common utils 
-* FIXED: Copyright comments to not reference year 
+* ADDED: Accept config for redis peer service attribute
+* ADDED: Move utf8 encoding to common utils
+* FIXED: Copyright comments to not reference year
 
 ### v0.10.1 / 2020-12-09
 
-* FIXED: Semantic conventions db.type -> db.system 
+* FIXED: Semantic conventions db.type -> db.system
 
 ### v0.10.0 / 2020-12-03
 
@@ -42,21 +46,21 @@
 
 ### v0.9.0 / 2020-11-27
 
-* BREAKING CHANGE: Add timeout for force_flush and shutdown 
+* BREAKING CHANGE: Add timeout for force_flush and shutdown
 
-* ADDED: Redis attribute propagation 
-* ADDED: Add timeout for force_flush and shutdown 
+* ADDED: Redis attribute propagation
+* ADDED: Add timeout for force_flush and shutdown
 
 ### v0.8.0 / 2020-10-27
 
-* BREAKING CHANGE: Remove 'canonical' from status codes 
+* BREAKING CHANGE: Remove 'canonical' from status codes
 
-* FIXED: Remove 'canonical' from status codes 
+* FIXED: Remove 'canonical' from status codes
 
 ### v0.7.0 / 2020-10-07
 
-* DOCS: Added redis documentation 
-* DOCS: Standardize toplevel docs structure and readme 
+* DOCS: Added redis documentation
+* DOCS: Standardize toplevel docs structure and readme
 
 ### v0.6.0 / 2020-09-10
